requirements.txt
instaloader
rich
openai-whisper
moviepy

